% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gex.r
\name{add_gex}
\alias{add_gex}
\title{Add gene expression values to variant result}
\usage{
add_gex(res, rec)
}
\arguments{
\item{res}{Annotated variants from `annotate_coding()`}

\item{rec}{Sample record from config file}
}
\value{
Annotated variants including gene counts and TPM
}
\description{
Add gene expression values to variant result
}
