% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_coding.r
\name{annotate_coding}
\alias{annotate_coding}
\title{Annotate VCF variants with coding changes}
\usage{
annotate_coding(vr, txdb, asm)
}
\arguments{
\item{vr}{A VRanges object with SNVs and small indels}

\item{txdb}{TxDb or EnsDb object}

\item{asm}{Genomic sequence BSGenome object}
}
\value{
A GRanges object with annotated variants
}
\description{
Annotate VCF variants with coding changes
}
\examples{
if (interactive()) {
    vcf = system.file("my_variants.vcf", package="pepitope")
    vr = readVcfAsVRanges(vcf)
    txdb = AnnotationHub::AnnotationHub()[["AH100643"]]
    asm = BSgenome.Hsapiens.NCBI.GRCh38::BSgenome.Hsapiens.NCBI.GRCh38
    annotate_coding(vr, txdb, asm)
}

}
