% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_fusions.r
\name{annotate_fusions}
\alias{annotate_fusions}
\title{Aggregate fusion VCFs into a table}
\usage{
annotate_fusions(vr, txdb, asm)
}
\arguments{
\item{vr}{A VRanges object with RNA fusions from readVcfAsRanges}

\item{txdb}{A transcription database, eg. AnnotationHub()[["AH100643"]]}

\item{asm}{A Genome sequence package object, eg. ::BSgenome.Hsapiens.NCBI.GRCh38}
}
\value{
A DataFrame objects with fusions
}
\description{
Aggregate fusion VCFs into a table
}
\examples{
if (interactive()) {
    vcf = system.file("my_fusions.vcf", package="pepitope")
    vr = readVcfAsVRanges(vcf)
    txdb = AnnotationHub::AnnotationHub()[["AH100643"]]
    asm = BSgenome.Hsapiens.NCBI.GRCh38::BSgenome.Hsapiens.NCBI.GRCh38
    annotate_fusions(vr, txdb, asm)
}

}
