% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_read_structure.r
\name{annotate_read_structure}
\alias{annotate_read_structure}
\title{Annotate barcode positions in source FASTQ reads}
\usage{
annotate_read_structure(fq, samples, all_constructs)
}
\arguments{
\item{fq}{Path to one FASTQ file}

\item{samples}{A sample sheet as `data.frame` in tsv format. Requires the
columns 'sample_id', 'patient', 'rep', 'origin', 'barcode'}

\item{all_constructs}{A named list of all construct libraries}
}
\value{
A `data.frame` with barcode matches per position
}
\description{
Annotate barcode positions in source FASTQ reads
}
