% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_calc.r
\name{calc_de}
\alias{calc_de}
\title{Calculate differential abundance of construct barcodes}
\usage{
calc_de(eset, cfg)
}
\arguments{
\item{eset}{A DESeq2 object}

\item{cfg}{outdated}
}
\description{
Calculate differential abundance of construct barcodes
}
