% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distr.r
\name{calc_representation}
\alias{calc_representation}
\title{Calculate ranked barcode representation across samples}
\usage{
calc_representation(lib_counts, bcs, meta)
}
\arguments{
\item{lib_counts}{A count matrix}

\item{bcs}{Barcodes}

\item{meta}{Sample sheet and read count summaries}
}
\description{
Calculate ranked barcode representation across samples
}
\keyword{internal}
