% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.r
\name{cds_by_break}
\alias{cds_by_break}
\title{Get a list of transcripts with their CDS exons overlapping the break}
\usage{
cds_by_break(gr, txdb, cds, type = "left")
}
\arguments{
\item{gr}{GenomicRanges object of break location}

\item{txdb}{A transcription database, eg. AnnotationHub()[["AH100643"]]}

\item{cds}{A list of exon coordinates by gene from `cdsBy(txdb)`}

\item{type}{Whether we want info for the 'left' or 'right' side of the break}
}
\value{
A named list of transcript GRanges objects with CDS exons
}
\description{
Get a list of transcripts with their CDS exons overlapping the break
}
\keyword{internal}
