% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_coding.r
\name{check_silent}
\alias{check_silent}
\title{Adjust offsets for variants that start/end silently (no AA change)}
\usage{
check_silent(ref, alt)
}
\arguments{
\item{ref}{Reference amino acid sequence}

\item{alt}{Alternative amino acid sequence}
}
\value{
An integer vector of offsets
}
\description{
Adjust offsets for variants that start/end silently (no AA change)
}
\keyword{internal}
