% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_fastq.r
\name{count_fastq}
\alias{count_fastq}
\title{Count barcodes directly from source FASTQ files}
\usage{
count_fastq(
  fq,
  samples,
  all_constructs,
  valid_barcodes,
  read_structures = "7B12M+T",
  reverse_complement = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fq}{Path to one FASTQ file}

\item{samples}{A sample sheet as \code{data.frame} in tsv format. Requires the
columns 'sample_id', 'patient', 'rep', 'origin', 'barcode'}

\item{all_constructs}{A named list of all construct libraries}

\item{valid_barcodes}{A character vector of all possible construct barcodes}

\item{read_structures}{A character string describing the FASTQ read structure. \code{B} segments
are matched against sample barcodes and \code{M} segments against construct barcodes.}

\item{reverse_complement}{Whether to count the reverse complement of the construct barcodes instead}

\item{verbose}{Whether to print progress messages (default: TRUE)}
}
\value{
A \code{SummarizedExperiment} object with counts and metadata
}
\description{
Count barcodes directly from source FASTQ files
}
