% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_demux.r
\name{demux_fq}
\alias{demux_fq}
\title{Use fqtk tool to demultiplex fastq files}
\usage{
demux_fq(fq, samples, read_structures)
}
\arguments{
\item{fq}{A path to the fastq file to demultiplex}

\item{samples}{A sample sheet as `data.frame` in tsv format. Requires the
columns 'sample_id', 'patient', 'rep', 'origin', 'barcode'}

\item{read_structures}{A character string describing the read structure}
}
\description{
Use fqtk tool to demultiplex fastq files
}
