% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.r
\name{.rs_annotate}
\alias{.rs_annotate}
\title{Annotate barcode positions in source FASTQ reads}
\usage{
.rs_annotate(fq, samples, all_constructs, nrec = 100000L)
}
\arguments{
\item{fq}{Path to one FASTQ file}

\item{samples}{A sample sheet as `data.frame` in tsv format. Requires the
columns 'sample_id', 'patient', 'rep', 'origin', 'barcode'}

\item{all_constructs}{A named list of all construct libraries}

\item{nrec}{Number of FASTQ records to inspect}
}
\value{
A `list` with barcode counts, reads, and inferred read structure
}
\description{
Annotate barcode positions in source FASTQ reads
}
\keyword{internal}
