% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.r
\name{.rs_parse}
\alias{.rs_parse}
\title{Parse a read structure}
\usage{
.rs_parse(read_structures)
}
\arguments{
\item{read_structures}{A character string describing the FASTQ read structure}
}
\value{
A `list` with sample and construct barcode ranges
}
\description{
Parse a read structure
}
\keyword{internal}
