% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{example_barcodes}
\alias{example_barcodes}
\title{Create deterministic example barcodes}
\usage{
example_barcodes(n, width = 12)
}
\arguments{
\item{n}{Number of barcodes to create}

\item{width}{Barcode width in nucleotides}
}
\value{
A character vector of DNA barcodes
}
\description{
Create deterministic example barcodes
}
\examples{
example_barcodes(4, width=4)

}
\keyword{internal}
