% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{example_fastq}
\alias{example_fastq}
\title{Simulate sequencing data and write them to a temporary FASTQ file}
\usage{
example_fastq(
  samples,
  peptide_sheets,
  target_reads = 1000,
  custom = TRUE,
  seed = 91651
)
}
\arguments{
\item{samples}{The .tsv or data.frame file containing sample information}

\item{peptide_sheets}{A list, each item containing construct information}

\item{target_reads}{How many reads to simulate on average}

\item{custom}{Whether to add custom modifications to founds}

\item{seed}{The random seed used for sampling the number of reads}
}
\value{
The path to the created FASTQ file
}
\description{
Simulate sequencing data and write them to a temporary FASTQ file
}
\examples{
samples = data.frame(sample_id="sample1", patient="pat1", rep="1",
    origin="library", barcode="GGG")
constructs = list(pat1=data.frame(gene_name="GENE1", mut_id="GENE1_A1V",
    pep_id="GENE1_A1V", pep_type="alt", tiled="ATGGCCGCC", barcode_1="AAAA"))
example_fastq(samples, constructs, target_reads=2, custom=FALSE)

}
\keyword{internal}
