% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{example_peptide_file}
\alias{example_peptide_file}
\title{Create a my_peptides.tsv in the inst directory}
\usage{
example_peptide_file()
}
\value{
Invisibly returns `NULL` after writing the example peptide file
}
\description{
Create a my_peptides.tsv in the inst directory
}
\keyword{internal}
