% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.r
\name{fusion}
\alias{fusion}
\title{Aggregate fusion VCFs into a table}
\usage{
fusion(vr, txdb, asm, min_reads = NULL, min_mates = NULL, min_tools = NULL)
}
\arguments{
\item{vr}{A VRanges object with RNA fusions from readVcfAsRanges}

\item{txdb}{A transcription database, eg. AnnotationHub()[["AH100643"]]}

\item{min_reads}{The minimum number of split read support for a fusion}

\item{min_mates}{The minimum number of linked read support for a fusion}

\item{min_tools}{The minimum number of tools that identify a fusion}
}
\value{
A DataFrame objects with fusions
}
\description{
Aggregate fusion VCFs into a table
}
