% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coding_seq.r
\name{get_coding_seq}
\alias{get_coding_seq}
\title{Get the alternative coding sequence of a variant or gene fusion}
\usage{
get_coding_seq(txdb, ...)
}
\arguments{
\item{txdb}{A transcription database, eg. AnnotationHub()[["AH100643"]]}

\item{...}{A named DNAStringSet objects where each row is translated consecutively}
}
\value{
A merged DNAStringSet object with the translated nucleotides
}
\description{
This takes into account any (newly introduced) STOP codons and UTR
readthroughs
}
\keyword{internal}
