% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proper_orf.r
\name{is_proper_orf}
\alias{is_proper_orf}
\title{Logical vector whether sequence has START and STOP codon}
\usage{
is_proper_orf(x)
}
\arguments{
\item{x}{A DNAStringSet or AAStringSet object}
}
\value{
A logical vector indicating whether each sequence has a proper ORF
}
\description{
Logical vector whether sequence has START and STOP codon
}
\keyword{internal}
