% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_calc.r
\name{make_clean}
\alias{make_clean}
\title{Clean up a differential abundance result}
\usage{
make_clean(res, sample, cap_fc = 3)
}
\arguments{
\item{res}{A DESeq2 results object}

\item{sample}{The sample name}

\item{cap_fc}{An absolute limit of the log2 fold change}
}
\value{
A cleaned results object
}
\description{
Clean up a differential abundance result
}
