% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{make_pep_table}
\alias{make_pep_table}
\title{Create a peptide .tsv in the inst directory}
\usage{
make_pep_table(outfile)
}
\arguments{
\item{outfile}{The file to save peptide table to}
}
\description{
Create a peptide .tsv in the inst directory
}
\keyword{internal}
