% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_peptides.r
\name{make_peptides}
\alias{make_peptides}
\title{Make a variants report as named list of tables}
\usage{
make_peptides(subs, fus = DataFrame())
}
\arguments{
\item{subs}{Variants within mutation context from `subset_context()`}

\item{fus}{Variants within fusion context from `subset_context_fusions()`}
}
\value{
A data.frame with cDNA and peptide sequences
}
\description{
Make a variants report as named list of tables
}
