% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.r
\name{make_report}
\alias{make_report}
\title{Make results report to save as xlsx sheets (full, filtered, peptides)}
\usage{
make_report(
  res,
  fname,
  min_cov = 2,
  min_af = 0.1,
  tile_size = 93,
  tile_ov = 45
)
}
\arguments{
\item{res}{A full results GRanges object from `annotate_coding()`}

\item{fname}{File name to save results to}

\item{min_cov}{Minimum number of reads to span the ALT allele}

\item{min_af}{Minimum allele frequency of the ALT allele}

\item{tile_size}{Oligo tiling size}

\item{tile_ov}{Oligo tiling overlap}
}
\description{
Make results report to save as xlsx sheets (full, filtered, peptides)
}
