% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_count.r
\name{merge_constructs}
\alias{merge_constructs}
\title{Bind the tables of all library annotations together}
\usage{
merge_constructs(all_constructs)
}
\arguments{
\item{all_constructs}{A named list of all construct libraries}
}
\value{
A `data.frame` with all constructs
}
\description{
Bind the tables of all library annotations together
}
\keyword{internal}
