% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_tile.r
\name{pep_tile}
\alias{pep_tile}
\title{Tile cDNA into peptide sequences}
\usage{
pep_tile(peptides, tile_size = 93, tile_ov = 45)
}
\arguments{
\item{peptides}{A `data.frame` with context-subset peptide/minigene data}

\item{tile_size}{Oligo tiling size}

\item{tile_ov}{Oligo tiling overlap}
}
\value{
A data.frame with tiled nucleotide and peptide sequences
}
\description{
Tile cDNA into peptide sequences
}
\examples{
peptides = data.frame(var_id="v1", mut_id="M1_A1V", gene_name="M1",
    gene_id="gene1", tx_id="tx1", pep_id="M1_A1V", cDNA="ATGGCCGCCGCC")
pep_tile(peptides, tile_size=9, tile_ov=3)

}
