% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepitope-package.r
\docType{package}
\name{pepitope}
\alias{pepitope}
\title{Pepitope: peptide epitopes from reference genome and variant (VCF) file}
\description{
Given a reference genome and VCF file, this package will provide
the upstream/downstream peptide context of a variant. It will generate a
summary report for protein-coding variants including the reference and
mutated allele, read coverage, amino acid sequence, and other information.
It can also be used to remove restriction sites from cDNA, alongside other
helper functions.
}
