% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_distr.r
\name{plot_distr}
\alias{plot_distr}
\title{Plot the read distribution across barcodes}
\usage{
plot_distr(reps)
}
\arguments{
\item{reps}{A `data.frame` from `calc_representation`}
}
\value{
A `ggplot2` object
}
\description{
Plot the read distribution across barcodes
}
