% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genomic_context.r
\name{plot_genomic_context}
\alias{plot_genomic_context}
\title{Genomic track plot for variant + transcripts + RNA-seq}
\usage{
plot_genomic_context(rec, gene_id, res, gene, txdb)
}
\description{
Genomic track plot for variant + transcripts + RNA-seq
}
