% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_count.r
\name{plot_read_count}
\alias{plot_read_count}
\alias{plot_reads}
\title{Plot the overall read counts}
\usage{
plot_read_count(dset, n_cutoff = 10)

plot_reads(dset)
}
\arguments{
\item{dset}{The `SummarizedExperiment` object from `count_fastq`}

\item{n_cutoff}{Minimum reads for counting a barcode as represented}
}
\value{
A patchwork object containing the read count summary plots
}
\description{
Plot the overall read counts
}
