% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reads.r
\name{plot_reads}
\alias{plot_reads}
\title{Plot the overall read counts}
\usage{
plot_reads(dset)
}
\arguments{
\item{dset}{The `SummarizedExperiment` object from `count_bc`}
}
\description{
Plot the overall read counts
}
