% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepitope-package.r
\name{readVcfAsVRanges}
\alias{readVcfAsVRanges}
\title{Re-export from VariantAnnotation package}
\usage{
readVcfAsVRanges(x, genome, param = ScanVcfParam(), use.names = FALSE, ...)
}
\description{
Re-export from VariantAnnotation package
}
