% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assay}
\alias{colData}
\alias{rowData}
\alias{genome}
\alias{seqinfo}
\alias{seqlevels}
\alias{seqlevelsStyle}
\alias{seqlevelsStyle<-}
\alias{readVcfAsVRanges}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GenomeInfoDb}{\code{\link[GenomeInfoDb:genome]{genome}}, \code{\link[GenomeInfoDb:seqinfo]{seqinfo}}, \code{\link[GenomeInfoDb:seqlevels]{seqlevels}}, \code{\link[GenomeInfoDb:seqlevelsStyle]{seqlevelsStyle()}}, \code{\link[GenomeInfoDb:seqlevelsStyle<-]{seqlevelsStyle<-()}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:assay]{assay()}}, \code{\link[SummarizedExperiment:colData]{colData()}}, \code{\link[SummarizedExperiment:rowData]{rowData()}}}

  \item{VariantAnnotation}{\code{\link[VariantAnnotation:readVcfAsVRanges]{readVcfAsVRanges()}}}
}}

