% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cutsite.r
\name{remove_cutsite}
\alias{remove_cutsite}
\title{Remove a Restriction Enzyme cut site but keep AA in a tiled peptide data.frame}
\usage{
remove_cutsite(pep, ...)
}
\arguments{
\item{pep}{A data.frame of tiled peptides}

\item{...}{Named argumennts of cut sites, e.g. `BbsI="GAAGAC"`}
}
\value{
A data.frame with replace nucleotides and number of replacements
}
\description{
Remove a Restriction Enzyme cut site but keep AA in a tiled peptide data.frame
}
\examples{
pep = data.frame(pep_id="p1", tiled="ATGGCCGCC")
remove_cutsite(pep, BbsI="GAAGAC")

}
