% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_tile.r
\name{report_tiles}
\alias{report_tiles}
\title{Merge different sources of tiles to one report}
\usage{
report_tiles(pep)
}
\arguments{
\item{...}{Tile sources}
}
\description{
Merge different sources of tiles to one report
}
