% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_variants.r
\name{report_variants}
\alias{report_variants}
\title{Make results report to save as xlsx sheets (full, filtered, peptides)}
\usage{
report_variants(res, fname, min_cov = 2, min_af = 0.1)
}
\arguments{
\item{res}{A full results GRanges object from `annotate_coding()`}

\item{fname}{File name to save results to}

\item{min_cov}{Minimum number of reads to span the ALT allele}

\item{min_af}{Minimum allele frequency of the ALT allele}
}
\description{
Make results report to save as xlsx sheets (full, filtered, peptides)
}
