% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_calc.r
\name{screen_calc}
\alias{screen_calc}
\title{Calculate differential abundance of construct barcodes}
\usage{
screen_calc(dset, comparisons)
}
\arguments{
\item{dset}{A `SummarizedExperiment` object from `count_fastq()`}

\item{comparisons}{A character vector of sample and reference condition, or list thereof}
}
\value{
A data.frame of DESeq2 results, or a named list of data.frames
}
\description{
Calculate differential abundance of construct barcodes
}
\examples{
if (interactive()) {
    screen_calc(dset, c("screen", "library"))
}

}
