% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{sim_fusions}
\alias{sim_fusions}
\title{Simulate gene fusions and return them as VRanges object}
\usage{
sim_fusions()
}
\description{
Simulate gene fusions and return them as VRanges object
}
\keyword{internal}
