% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{sim_variants}
\alias{sim_variants}
\title{Simulate mutations and return them as VRanges object}
\usage{
sim_variants()
}
\description{
Simulate mutations and return them as VRanges object
}
\keyword{internal}
