% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{sim_vcf}
\alias{sim_vcf}
\title{Simulate mutations and write them into a VCF file}
\usage{
sim_vcf(fname)
}
\arguments{
\item{fname}{File name of the VCF to save}
}
\description{
Simulate mutations and write them into a VCF file
}
\keyword{internal}
