% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_context.r
\name{subset_context}
\alias{subset_context}
\title{Subset nucleotide/protein sequences to codon +/- 45 bp context}
\usage{
subset_context(codv, ctx_codons = 15)
}
\arguments{
\item{codv}{Annotated variants from `annotate_coding()`}

\item{ctx_codons}{How many flanking codons each to include in the context}
}
\value{
GRanges object with sequence information of only context
}
\description{
Subset nucleotide/protein sequences to codon +/- 45 bp context
}
