% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_context_fusion.r
\name{subset_context_fusion}
\alias{subset_context_fusion}
\title{Subset the peptide context for gene fusions}
\usage{
subset_context_fusion(res, ctx_codons)
}
\arguments{
\item{res}{A DataFrame object from `fusions`}

\item{ctx_codons}{How many flanking codons each to include in the context}
}
\value{
A DataFrame object of gene fusions
}
\description{
Subset the peptide context for gene fusions
}
