% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.r
\name{tx_by_break}
\alias{tx_by_break}
\title{Get transcript incl coords left or right of break}
\usage{
tx_by_break(gr, asm, txdb, tx, cds, type = "left")
}
\arguments{
\item{gr}{GenomicRanges object of break location}

\item{asm}{A Genome sequence package object, eg. ::BSgenome.Hsapiens.NCBI.GRCh38}

\item{txdb}{A transcription database, eg. AnnotationHub()[["AH100643"]]}

\item{tx}{A list of transcripts obtained from `transcripts(txdb)`}

\item{cds}{A list of exon coordinates by gene from `cdsBy(txdb)`}

\item{type}{Whether we want info for the 'left' or 'right' side of the break}
}
\value{
A DataFrame with sequence information
}
\description{
Get transcript incl coords left or right of break
}
\keyword{internal}
