% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.r
\name{tx_combine_breaks}
\alias{tx_combine_breaks}
\title{Combine break info from each possible left and right side transcript}
\usage{
tx_combine_breaks(vr, left, right)
}
\arguments{
\item{vr}{A VRanges object with RNA fusions from `readVcfAsVRanges`}

\item{left}{List of DataFrame objects containing the 5' of the fusion}

\item{right}{List of DataFrame objects containing the 3' of the fusion}
}
\value{
A DataFrame with fusion coordinates and sequence information
}
\description{
Combine break info from each possible left and right side transcript
}
\keyword{internal}
