% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_filter.r
\name{variant_filter}
\alias{variant_filter}
\title{Make results report to save as xlsx sheets (full, filtered, peptides)}
\usage{
variant_filter(res, min_cov = 2, min_af = 0.1)
}
\arguments{
\item{res}{A full results GRanges object from `annotate_coding()`}

\item{min_cov}{Minimum number of reads to span the ALT allele}

\item{min_af}{Minimum allele frequency of the ALT allele}
}
\description{
Make results report to save as xlsx sheets (full, filtered, peptides)
}
