% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_coding.r
\name{annotate_coding}
\alias{annotate_coding}
\title{Annotate VCF variants with coding changes}
\usage{
annotate_coding(vr, txdb, asm, filter_variants = FALSE)
}
\arguments{
\item{vr}{A VRanges object with SNVs and small indels}

\item{txdb}{TxDb or EnsDb object}

\item{asm}{Genomic sequence BSGenome object}

\item{filter_variants}{Apply soft filter matrix to the variants}
}
\value{
A GRanges object with annotated variants
}
\description{
Annotate VCF variants with coding changes
}
