% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_coding.r
\name{annotate_coding}
\alias{annotate_coding}
\title{Annotate VCF variants with coding changes}
\usage{
annotate_coding(rec, txdb, asm, tx_coding, tumor_cov = "tumor_DNA")
}
\arguments{
\item{rec}{A sample record object}

\item{txdb}{Txdb object}

\item{asm}{Genomic sequence BSGenome object}

\item{tx_coding}{Character vector of ENST0000 IDs that are protein coding}

\item{tumor_cov}{Part of column name to get ref/alt coverage (regex)}
}
\value{
A GRanges object with annotated variants
}
\description{
Annotate VCF variants with coding changes
}
