% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_count.r
\name{count_bc}
\alias{count_bc}
\title{Count barcodes using guide-counter}
\usage{
count_bc(tdir, all_constructs, valid_barcodes, reverse_complement = FALSE)
}
\arguments{
\item{tdir}{Path to the directory with demultiplexed FASTQ files}

\item{all_constructs}{A named list of all construct libraries}

\item{valid_barcodes}{A character vector of all possible barcodes}

\item{reverse_complement}{Whether to count the reverse complement of the barcodes instead}
}
\value{
A list with the data.frame meta and matrix counts
}
\description{
Count barcodes using guide-counter
}
