% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_count.r
\name{count_bc}
\alias{count_bc}
\title{Count barcodes using guide-counter}
\usage{
count_bc(tdir, lib)
}
\arguments{
\item{tdir}{Path to the directory with demultiplexed FASTQ files}

\item{lib}{Path to the barcode library file}
}
\value{
A list with the data.frame meta and matrix counts
}
\description{
Count barcodes using guide-counter
}
