% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_count.r
\name{count_external}
\alias{count_external}
\title{Use `guide-counter` via a system call to actually count}
\usage{
count_external(tdir, valid_barcodes, reverse_complement)
}
\arguments{
\item{tdir}{Path to the directory with demultiplexed FASTQ files}

\item{valid_barcodes}{A character vector of all possible barcodes}

\item{reverse_complement}{Whether to count the reverse complement of the barcodes instead}
}
\value{
A list with the data.frame meta and matrix counts
}
\description{
Use `guide-counter` via a system call to actually count
}
\keyword{internal}
