% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.r
\name{.rs_format}
\alias{.rs_format}
\title{Format a read structure from barcode position counts}
\usage{
.rs_format(counts, sample_width, construct_width)
}
\arguments{
\item{counts}{A `data.frame` with `B`, `B<`, `M`, and `M<` count columns}

\item{sample_width}{Width of the sample barcode}

\item{construct_width}{Width of the construct barcode}
}
\value{
A read structure string
}
\description{
Format a read structure from barcode position counts
}
\keyword{internal}
