% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{example_fastq}
\alias{example_fastq}
\title{Simulate sequencing data and write them to a temporary FASTQ file}
\usage{
example_fastq(samples, peptide_sheets, target_reads = 1000)
}
\arguments{
\item{target_reads}{How many reads to simulate on average}

\item{sample_sheet}{The .tsv or data.frame file containing sample information}

\item{peptide_sheet}{A list, each item containing construct information}
}
\value{
The path to the created FASTQ file
}
\description{
Simulate sequencing data and write them to a temporary FASTQ file
}
