% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{example_peptides}
\alias{example_peptides}
\title{Create example peptide sheets for multiple samples}
\usage{
example_peptides(valid_barcodes)
}
\arguments{
\item{valid_barcodes}{A character vector of valid barcodes}
}
\value{
A named list of peptide/minigene constructs with barcodes
}
\description{
Create example peptide sheets for multiple samples
}
\examples{
bases = expand.grid(rep(list(c("A", "C", "G", "T")), 4))
valid_barcodes = apply(bases, 1, paste0, collapse="")
if (interactive()) {
    example_peptides(valid_barcodes)
}

}
\keyword{internal}
