% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.r
\name{filter_variants}
\alias{filter_variants}
\title{Make results report to save as xlsx sheets (full, filtered, peptides)}
\usage{
filter_variants(
  vr,
  ...,
  min_cov = 2,
  min_af = 0.05,
  pass = TRUE,
  sample = NULL,
  chrs = NULL
)
}
\arguments{
\item{vr}{A VRanges object from `readVcfAsVRanges`}

\item{...}{Force filters by name (ignored)}

\item{min_cov}{Minimum number of reads to span the ALT allele}

\item{min_af}{Minimum allele frequency of the ALT allele}

\item{pass}{Whether to only include softFilterMatrix PASS}

\item{sample}{Only include if in `sampleNames(vr)` (required if more than one present)}

\item{chrs}{Either "default" or a character vector of chromosome names}
}
\description{
Make results report to save as xlsx sheets (full, filtered, peptides)
}
