% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.r
\name{fusion}
\alias{fusion}
\title{Aggregate fusion VCFs into a table}
\usage{
fusion(
  vr,
  txdb,
  asm,
  min_reads = NULL,
  min_pairs = NULL,
  min_tools = NULL,
  ctx_codons = 15
)
}
\arguments{
\item{vr}{A VRanges object with RNA fusions from readVcfAsRanges}

\item{txdb}{A transcription database, eg. AnnotationHub()[["AH100643"]]}

\item{min_reads}{The minimum number of split read support for a fusion}

\item{min_pairs}{The minimum number of linked read support for a fusion}

\item{min_tools}{The minimum number of tools that identify a fusion}

\item{ctx_codons}{Number of codonds for sequence context}
}
\value{
A DataFrame objects with fusions
}
\description{
Aggregate fusion VCFs into a table
}
