% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.r
\name{fusion}
\alias{fusion}
\title{Aggregate fusion VCFs into a table}
\usage{
fusion(vr, txdb, asm, filter_fusions = FALSE)
}
\arguments{
\item{vr}{A VRanges object with RNA fusions from readVcfAsRanges}

\item{txdb}{A transcription database, eg. AnnotationHub()[["AH100643"]]}

\item{filter_fusions}{Whether to only consider fusions with 2 tools, 2 reads}
}
\value{
A DataFrame objects with fusions
}
\description{
Aggregate fusion VCFs into a table
}
