% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.r
\name{make_report}
\alias{make_report}
\title{Make a variants report as named list of tables}
\usage{
make_report(res, subs, fus, ...)
}
\arguments{
\item{res}{Variant results}

\item{subs}{Filtered variant results}

\item{fus}{Fusion results}

\item{...}{Parameters passed to `pep_tile` (eg. tile_size, tile_ov)}
}
\description{
Make a variants report as named list of tables
}
