% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.r
\name{make_report}
\alias{make_report}
\title{Make a variants report as named list of tables}
\usage{
make_report(vars, fus = DataFrame(), ..., ctx_codons)
}
\arguments{
\item{vars}{Variant results}

\item{fus}{Fusion results}

\item{...}{Parameters passed to `pep_tile` (eg. tile_size, tile_ov)}

\item{ctx_codons}{Number of codonds for sequence context}
}
\description{
Make a variants report as named list of tables
}
