% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.r
\name{make_report}
\alias{make_report}
\title{Make a variants report as named list of tables}
\usage{
make_report(vars, subs, fus = DataFrame(), tiled)
}
\arguments{
\item{vars}{Variant results from `annotate_coding()`}

\item{subs}{Variants within mutation context from `subset_context()`}

\item{fus}{Variants within fusion context from `subset_context_fusions()`}

\item{tiled}{A data.frame of the tiled peptide sequences}
}
\description{
Make a variants report as named list of tables
}
