% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_tile.r
\name{pep_tile}
\alias{pep_tile}
\title{Tile cDNA into peptide sequences}
\usage{
pep_tile(subs, tile_size = 93, tile_ov = 45)
}
\arguments{
\item{subs}{A GRanges object of context-subset protein-coding variants}

\item{tile_size}{Oligo tiling size}

\item{tile_ov}{Oligo tiling overlap}
}
\description{
Tile cDNA into peptide sequences
}
