% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barcode_overlap.r
\name{plot_barcode_overlap}
\alias{plot_barcode_overlap}
\title{Plot barcode overlap between different samples}
\usage{
plot_barcode_overlap(all_constructs, valid_barcodes)
}
\arguments{
\item{all_constructs}{A named list of all constructs}

\item{valid_barcodes}{A character vector of possible barcodes (optional)}
}
\value{
A ggplot2 object
}
\description{
Plot barcode overlap between different samples
}
