% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distr.r
\name{plot_distr}
\alias{plot_distr}
\title{Plot the read distribution across barcodes}
\usage{
plot_distr(dset)
}
\arguments{
\item{dset}{The `SummarizedExperiment` object from `count_bc`}
}
\value{
A `ggplot2` object
}
\description{
Plot the read distribution across barcodes
}
