% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genomic_context.r
\name{plot_genomic_context}
\alias{plot_genomic_context}
\title{Genomic track plot for variant + transcripts + RNA-seq}
\usage{
plot_genomic_context(rec, gene_id, res, gene, txdb)
}
\arguments{
\item{rec}{A record}

\item{gene_id}{A gene id}

\item{res}{results object}

\item{gene}{gene name}

\item{txdb}{txdb object}
}
\value{
A plot
}
\description{
Genomic track plot for variant + transcripts + RNA-seq
}
\keyword{internal}
