% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_structure.r
\name{plot_read_structure}
\alias{plot_read_structure}
\title{Plot annotated read structure examples}
\usage{
plot_read_structure(fq, samples, all_constructs, nrec = 100000L)
}
\arguments{
\item{fq}{Path to one FASTQ file}

\item{samples}{A sample sheet as `data.frame` in tsv format. Requires the
columns 'sample_id', 'patient', 'rep', 'origin', 'barcode'}

\item{all_constructs}{A named list of all construct libraries}

\item{nrec}{Number of FASTQ records to inspect}
}
\value{
A `ggplot2` object
}
\description{
Plot annotated read structure examples
}
\examples{
if (interactive()) {
    samples = data.frame(sample_id="sample1", patient="pat1", rep="1",
        origin="library", barcode="GGG")
    constructs = list(pat1=data.frame(gene_name="GENE1", mut_id="GENE1_A1V",
        pep_id="GENE1_A1V", pep_type="alt", tiled="ATGGCCGCC", barcode_1="AAAA"))
    fq = tempfile(fileext=".fq")
    writeLines(c("@r1", "GGGAAAA", "+", "IIIIIII", "@r2", "GGGAAAA", "+", "IIIIIII"), fq)
    plot_read_structure(fq, samples, constructs, nrec=10)
}

}
