% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_screen.r
\name{plot_screen}
\alias{plot_screen}
\title{Plot screen results}
\usage{
plot_screen(res, sample = NULL, links = TRUE, labs = TRUE, cap_fc = 8)
}
\arguments{
\item{res}{A results `data.frame` from `screen_calc()`}

\item{sample}{Which library to plot (default: all)}

\item{links}{Whether to draw arrows between ref and significant alt peptides}

\item{labs}{Whether to label genes in less dense areas}

\item{cap_fc}{Maximum amount of fold-change to limit values to}
}
\value{
A `ggplot2` object of the differential expression results
}
\description{
Plot screen results
}
\examples{
if (interactive()) {
    res = data.frame(baseMean=c(20, 30), log2FoldChange=c(-1, 1), padj=c(0.2, 0.05),
        gene_name=c("GENE1", "GENE1"), pep_type=c("ref", "alt"),
        bc_type=c("pat1", "pat1"), mut_id=c("GENE1_A1", "GENE1_A1"))
    plot_screen(res, links=FALSE, labs=FALSE)
}

}
