% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepitope-package.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{readVcfAsVRanges}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{VariantAnnotation}{\code{\link[VariantAnnotation:VRanges-class]{readVcfAsVRanges}}}
}}

