% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepitope-package.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{readVcfAsVRanges}
\alias{colData}
\alias{rowData}
\alias{assay}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}}

  \item{VariantAnnotation}{\code{\link[VariantAnnotation:VRanges-class]{readVcfAsVRanges}}}
}}

