% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepitope-package.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{readVcfAsVRanges}
\alias{colData}
\alias{rowData}
\alias{assay}
\alias{seqinfo}
\alias{genome}
\alias{seqlevels}
\alias{seqlevelsStyle}
\alias{seqlevelsStyle<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GenomeInfoDb}{\code{\link[GenomeInfoDb:seqinfo]{genome}}, \code{\link[GenomeInfoDb]{seqinfo}}, \code{\link[GenomeInfoDb:seqinfo]{seqlevels}}, \code{\link[GenomeInfoDb]{seqlevelsStyle}}, \code{\link[GenomeInfoDb:seqlevelsStyle]{seqlevelsStyle<-}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}}

  \item{VariantAnnotation}{\code{\link[VariantAnnotation:VRanges-class]{readVcfAsVRanges}}}
}}

