% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cutsite.r
\name{remove_cutsite}
\alias{remove_cutsite}
\title{Remove a Restriction Enzyme cut site but keep AA}
\usage{
remove_cutsite(nuc, site, seed = NULL)
}
\arguments{
\item{nuc}{cDNA nucleotide string}

\item{site}{Recognition site to be replaced (fwd+rev comp)}

\item{seed}{Set random seed to select same changes on multiple runs}
}
\value{
cDNA with minimal changes to no longer contain the cut site
}
\description{
Remove a Restriction Enzyme cut site but keep AA
}
