% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_calc.r
\name{screen_calc}
\alias{screen_calc}
\title{Calculate differential abundance of construct barcodes}
\usage{
screen_calc(eset, cfg)
}
\arguments{
\item{eset}{A DESeq2 object}

\item{cfg}{outdated}
}
\description{
Calculate differential abundance of construct barcodes
}
