% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_calc.r
\name{screen_calc}
\alias{screen_calc}
\title{Calculate differential abundance of construct barcodes}
\usage{
screen_calc(dset, comparisons)
}
\arguments{
\item{dset}{A `SummarizedExperiment` object from `count_bc()`}

\item{comparisons}{A character vector of sample and reference condition, or list thereof}
}
\description{
Calculate differential abundance of construct barcodes
}
