% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{sim_fastq}
\alias{sim_fastq}
\title{Simulate sequencing data and write them to a FASTQ file}
\usage{
sim_fastq(sample_sheet, peptide_sheet, target_reads = 1000)
}
\arguments{
\item{sample_sheet}{The .tsv file containing sample information}

\item{peptide_sheet}{The .tsv file containing construct information}

\item{target_reads}{How many reads to simulate on average}
}
\description{
Simulate sequencing data and write them to a FASTQ file
}
